Initial population
This is version 1.1 of the website as delivered by Scott Peacock of SPTechnology, together with changes from Bob Summerwill to import new text and start to integrate that into MainVue.js.

Bob's comments:

Added well-nested text strings to both the English and Chinese JSON files.

This is the latest text from my draft website, but added in a form which also conveys the internal structure, so it is more maintainable.

The Chinese JSON is just a copy of the English file for now, and will need all of its strings to be translated.

As of this commit, these strings are not being used in HomeView.vue yet.  That is phase 2, and will be a pain because all that structure needs to be redone.

These translations will not include any hyperlinks yet, because those are not cleanly modelled in the existing Vue file.

-----

Switched "funding" section to use the new strings.
The changes themselves were not too hard, but it's hard working with the raw HTML which is very verbose, with lots of cut-and-paste and no comments or clear structure.

-----

Rewrote the "Overview" section using the new About text.   The hyperlinks are lost in the process, but can be added back later.

Added various comments, noting where the source file maps into the webpage, and some note on anchors in the wrong place, presumably to make the navigation look better, but why is it broken in the first place?

Removed "apply" chunk of text because it is redundant in this website, which has the "Hero text" instead.

Added "shortened" strings for the navigation view.

The HTML boilerplate is huge and ugly cut-and-paste, with formatting coming from both div and p.

Not good abstraction at all and hard to maintain.