Messing around
The translation supports nested strings, and I can do a horrible substitution like this with hardcoded URLs in Paragraph, but they all end up inside 'p' elements where we actually need to be generating anchors.

Looks to me like we are going to need to be using Vue components in a strong form for this to be possible, not just using Vue components to aggregate raw HTML with string substitutions.