Switch in the new logo on main page and as favicon
It's a bit smaller than I would like on the main page, and also using the text below option rather than beside.

Having too wide an icon is screwing up the other headinfgs, with what appears to be loads of padding between the icon and to the other nav headings.